refactor: organize parsers in a way that are scalable for future added models
